+++ src/features/list/hooks/useOompaList.tsx
@@ -3,47 +3,10 @@ import { getOompaList } from '../services';
 import { CACHE_TIME, QUERY_KEY_LIST } from '../../../utils/constants';
 import useOompaListActions from './useOompaListActions';
 import { useEffect } from 'react';
-import { IOompaList, IOompaListItem, IOompaListWithStamp } from '../interfaces/oompaList';
+import { IOompaListItem, IOompaListWithStamp } from '../interfaces/oompaList';
 import { isDataExpired } from '../../../utils';
 import { useAppSelector } from '../../../hooks';
-
-const isValidKey = (key: keyof IOompaListItem) =>
-  key === 'id' ||
-  key === 'first_name' ||
-  key === 'gender' ||
-  key === 'profession' ||
-  key === 'image';
-
-const isValidValue = (key: keyof IOompaListItem, item: IOompaListItem) => {
-  switch (key) {
-    case 'gender':
-      return item.gender === 'M' || item.gender === 'F' || item.gender === '';
-    default:
-      return typeof item[key] === 'string';
-  }
-};
-
-const isValidOompaListItem = (item: IOompaListItem) => {
-  for (const key in item) {
-    if (!isValidKey(key as keyof IOompaListItem)) {
-      console.error(`Invalid Oompa List Item Property: ${key}`);
-      return;
-    } else if (!isValidValue(key as keyof IOompaListItem, item)) {
-      console.error(`Invalid Oompa List Item Value: ${key} : ${item[key as keyof IOompaListItem]}`);
-      return;
-    } else {
-    }
-  }
-  return true;
-};
-
-const isValidOompaList = (data: IOompaList) => {
-  return (
-    typeof data.current_page === 'number' &&
-    Array.isArray(data.items) &&
-    data.items.every(isValidOompaListItem)
-  );
-};
+import { isValidOompaList } from '../../detail/hooks/utils';
 
 const useOompaList = () => {
   const { setOompaList, setOompaListStamp } = useOompaListActions();
